Mock Backend Service for Testing API Calls from Firebase (#73)
Creates a mock express app that can be used for integration tests. This results in two distinct test workflows: One using the mocked API (which should be updated by the companion-app developer) and one referencing the latest version of the `data-services` API pulled from the `master` branch. The latter will fail independently in the case that the backend repo is not up to date.